+++ src/nwb_datajoint/decoding/clusterless.py
@@ -15,8 +15,8 @@ from nwb_datajoint.common.common_interval import IntervalList
 from nwb_datajoint.common.common_nwbfile import AnalysisNwbfile
 from nwb_datajoint.common.common_position import IntervalPositionInfo
 from nwb_datajoint.common.dj_helper_fn import fetch_nwb
-from nwb_datajoint.decoding.core import (convert_classes_to_dict,
-                                         restore_classes)
+from nwb_datajoint.decoding.dj_decoder_conversion import (
+    convert_classes_to_dict, restore_classes)
 from nwb_datajoint.spikesorting.spikesorting_curation import (
     CuratedSpikeSorting, Curation)
 from replay_trajectory_classification.classifier import (
