+++ src/nwb_datajoint/decoding/sorted_spikes.py
@@ -6,8 +6,8 @@ import pandas as pd
 from nwb_datajoint.common.common_interval import IntervalList
 from nwb_datajoint.common.common_nwbfile import AnalysisNwbfile
 from nwb_datajoint.common.dj_helper_fn import fetch_nwb
-from nwb_datajoint.decoding.core import (convert_classes_to_dict,
-                                         restore_classes)
+from nwb_datajoint.decoding.dj_decoder_conversion import (
+    convert_classes_to_dict, restore_classes)
 from nwb_datajoint.spikesorting.spikesorting_curation import \
     CuratedSpikeSorting
 from replay_trajectory_classification.classifier import (
